Extract Remote::BinaryFile
