+++ spec/lib/everything/blog/remote/binary_file_spec.rb
@@ -1,0 +1,5 @@
+require './lib/everything/blog/remote/binary_file'
+
+describe Everything::Blog::Remote::BinaryFile do
+end
+
